DeleteUser moved to user controller, test written
